requirements.txt
torch~=2.4.0
pillow~=11.3.0
torchvision~=0.19.0
tqdm~=4.67.1
numpy~=2.0.2
ftfy~=6.3.1
regex~=2025.7.34
scipy~=1.13.1
scikit-learn~=1.6.1
libmtl~=1.1.5
qpsolvers~=4.7.0
cvxpy~=1.7.1
networkx~=3.2.1
setuptools~=75.8.2
opencv-python~=4.12.0
einops~=0.8.1
seaborn~=0.13.2
matplotlib~=3.9.4
pandas~=2.3.1
timm~=1.0.19